Move Task class and make all possible Tasks use it.
