+++ src/custom-resources/index.ts
@@ -1,4 +1,6 @@
+export * from './task';
 export * from './handlers/lambda_custom_resource_handler';
 export * from './custom_resources_utilities';
+export * from './run_resource_always';
 export * from './encode_resource';
 export * from './lambda_custom_resource';
